Adjust for part2 and optimize
Part2 requires to duplicate each line 5 times.
Doing that naivly, the program does not terminate within reasonable
time.

The main performance improvement here is to use a HashMap to cache
results, so when we iterate the tree and hit a state that we had
previously, we do not re-compute everything.

Furthermore, I also skip sequences of operational springs. This did not
seem to improve performance much but still should give some
improvements.